refactor(parsers): move top_albums to subpackage
